Refactor `oct-cloud`
- Split `aws.rs` into `client.rs` and `resource.rs` in `oct-cloud/src/aws` module

Closes #135
