+++ crates/oct-cloud/src/aws/mod.rs
@@ -1,0 +1,3 @@
+pub mod resource;
+
+mod client;
