+++ crates/oct-cloud/src/lib.rs
@@ -1,2 +1,4 @@
-pub mod aws;
+pub mod resource;
 pub mod state;
+
+pub mod aws;
